refactor: move the NavigationGraph to the navgraphs package
